test(llc): migrate channel update message tests to stream_chat_test
