+++ packages/stream_chat/test/src/client/channel/channel_test.dart
@@ -1177,929 +1177,6 @@ void main() {
       );
     });
 
-    group('`.updateMessage`', () {
-      test('should work fine', () async {
-        final message = Message(
-          id: 'test-message-id',
-          state: MessageState.sent,
-        );
-
-        final updateMessageResponse = UpdateMessageResponse()..message = message;
-
-        when(
-          () => client.updateMessage(any(that: isSameMessageAs(message))),
-        ).thenAnswer((_) async => updateMessageResponse);
-
-        expectLater(
-          // skipping first seed message list -> [] messages
-          channel.state?.messagesStream.skip(1),
-          emitsInOrder([
-            [
-              isSameMessageAs(
-                message.copyWith(state: MessageState.updating),
-                matchMessageState: true,
-              ),
-            ],
-            [
-              isSameMessageAs(
-                message.copyWith(state: MessageState.updated),
-                matchMessageState: true,
-              ),
-            ],
-          ]),
-        );
-
-        final res = await channel.updateMessage(message);
-
-        expect(res, isNotNull);
-        expect(res.message.id, message.id);
-
-        verify(
-          () => client.updateMessage(
-            any(that: isSameMessageAs(message)),
-          ),
-        ).called(1);
-      });
-
-      test('with attachments should work just fine', () async {
-        final attachments = List.generate(
-          3,
-          (index) => Attachment(
-            id: 'test-attachment-id-$index',
-            type: index.isEven ? 'image' : 'file',
-            file: AttachmentFile(size: index * 33, path: 'test-file-path'),
-          ),
-        );
-
-        final message = Message(
-          id: 'test-message-id',
-          attachments: attachments,
-        );
-
-        final sendImageResponse = SendImageResponse()..file = 'test-image-url';
-        final sendFileResponse = SendFileResponse()..file = 'test-file-url';
-
-        when(
-          () => client.sendImage(
-            any(),
-            channelId,
-            channelType,
-            onSendProgress: any(named: 'onSendProgress'),
-            cancelToken: any(named: 'cancelToken'),
-            extraData: any(named: 'extraData'),
-          ),
-        ).thenAnswer((_) async => sendImageResponse);
-
-        when(
-          () => client.sendFile(
-            any(),
-            channelId,
-            channelType,
-            onSendProgress: any(named: 'onSendProgress'),
-            cancelToken: any(named: 'cancelToken'),
-            extraData: any(named: 'extraData'),
-          ),
-        ).thenAnswer((_) async => sendFileResponse);
-
-        when(
-          () => client.updateMessage(
-            any(that: isSameMessageAs(message)),
-          ),
-        ).thenAnswer(
-          (_) async => UpdateMessageResponse()
-            ..message = message.copyWith(
-              state: MessageState.sent,
-              attachments: attachments
-                  .map((it) => it.copyWith(uploadState: const UploadState.success()))
-                  .toList(growable: false),
-            ),
-        );
-
-        expectLater(
-          // skipping first seed message list -> [] messages
-          channel.state?.messagesStream.skip(1),
-          emitsInOrder(
-            [
-              // preparing attachments to upload
-              [
-                isSameMessageAs(
-                  message.copyWith(
-                    state: MessageState.updating,
-                    attachments: [...attachments.map((it) => it.copyWith(uploadState: const UploadState.preparing()))],
-                  ),
-                  matchMessageState: true,
-                  matchAttachments: true,
-                  matchAttachmentsUploadState: true,
-                ),
-              ],
-              // 0th attachment is successfully uploaded
-              [
-                isSameMessageAs(
-                  message.copyWith(
-                    state: MessageState.updating,
-                    attachments: [...attachments]
-                      ..[0] = attachments[0].copyWith(
-                        uploadState: const UploadState.success(),
-                      ),
-                  ),
-                  matchMessageState: true,
-                  matchAttachments: true,
-                  matchAttachmentsUploadState: true,
-                ),
-              ],
-              // 0th and 1st attachment is successfully uploaded
-              [
-                isSameMessageAs(
-                  message.copyWith(
-                    state: MessageState.updating,
-                    attachments: [...attachments]
-                      ..[0] = attachments[0].copyWith(
-                        uploadState: const UploadState.success(),
-                      )
-                      ..[1] = attachments[1].copyWith(
-                        uploadState: const UploadState.success(),
-                      ),
-                  ),
-                  matchMessageState: true,
-                  matchAttachments: true,
-                  matchAttachmentsUploadState: true,
-                ),
-              ],
-              // all the attachments are successfully uploaded
-              [
-                isSameMessageAs(
-                  message.copyWith(
-                    state: MessageState.updating,
-                    attachments: [...attachments.map((it) => it.copyWith(uploadState: const UploadState.success()))],
-                  ),
-                  matchMessageState: true,
-                  matchAttachments: true,
-                  matchAttachmentsUploadState: true,
-                ),
-              ],
-              [
-                isSameMessageAs(
-                  message.copyWith(
-                    state: MessageState.updated,
-                    attachments: [...attachments.map((it) => it.copyWith(uploadState: const UploadState.success()))],
-                  ),
-                  matchMessageState: true,
-                  matchAttachments: true,
-                  matchAttachmentsUploadState: true,
-                ),
-              ],
-            ],
-          ),
-        );
-
-        final res = await channel.updateMessage(message);
-
-        expect(res, isNotNull);
-        expect(res.message.id, message.id);
-        expect(res.message.attachments.length, message.attachments.length);
-        expect(
-          res.message.attachments.every(
-            (it) => it.uploadState == const UploadState.success(),
-          ),
-          isTrue,
-        );
-
-        verify(
-          () => client.sendImage(
-            any(),
-            channelId,
-            channelType,
-            onSendProgress: any(named: 'onSendProgress'),
-            cancelToken: any(named: 'cancelToken'),
-            extraData: any(named: 'extraData'),
-          ),
-        ).called(2);
-
-        verify(
-          () => client.sendFile(
-            any(),
-            channelId,
-            channelType,
-            onSendProgress: any(named: 'onSendProgress'),
-            cancelToken: any(named: 'cancelToken'),
-            extraData: any(named: 'extraData'),
-          ),
-        ).called(1);
-
-        verify(
-          () => client.updateMessage(
-            any(that: isSameMessageAs(message)),
-          ),
-        ).called(1);
-      });
-
-      test('should update message state even when error is not StreamChatNetworkError', () async {
-        final message = Message(
-          id: 'test-message-id-error-1',
-          state: MessageState.sent,
-        );
-
-        when(
-          () => client.updateMessage(
-            any(that: isSameMessageAs(message)),
-            skipEnrichUrl: true,
-          ),
-        ).thenThrow(ArgumentError('Invalid argument'));
-
-        expectLater(
-          // skipping first seed message list -> [] messages
-          channel.state?.messagesStream.skip(1),
-          emitsInOrder([
-            [
-              isSameMessageAs(
-                message.copyWith(state: MessageState.updating),
-                matchMessageState: true,
-              ),
-            ],
-            [
-              isSameMessageAs(
-                message.copyWith(
-                  state: MessageState.updatingFailed(
-                    skipPush: false,
-                    skipEnrichUrl: true,
-                  ),
-                ),
-                matchMessageState: true,
-              ),
-            ],
-          ]),
-        );
-
-        try {
-          await channel.updateMessage(message, skipEnrichUrl: true);
-        } catch (e) {
-          expect(e, isA<ArgumentError>());
-        }
-      });
-
-      test(
-        'should add message to retry queue when retriable StreamChatNetworkError occurs with skipPush: false, skipEnrichUrl: true',
-        () async {
-          final message = Message(
-            id: 'test-message-id-retry-1',
-            state: MessageState.sent,
-          );
-
-          // Create a retriable error (data == null)
-          when(
-            () => client.updateMessage(
-              any(that: isSameMessageAs(message)),
-              skipEnrichUrl: true,
-            ),
-          ).thenThrow(
-            StreamChatNetworkError.raw(
-              code: ChatErrorCode.requestTimeout.code,
-              message: 'Request timed out',
-            ),
-          );
-
-          expectLater(
-            // skipping first seed message list -> [] messages
-            channel.state?.messagesStream.skip(1),
-            emitsInOrder([
-              [
-                isSameMessageAs(
-                  message.copyWith(state: MessageState.updating),
-                  matchMessageState: true,
-                ),
-              ],
-              [
-                isSameMessageAs(
-                  message.copyWith(
-                    state: MessageState.updatingFailed(
-                      skipPush: false,
-                      skipEnrichUrl: true,
-                    ),
-                  ),
-                  matchMessageState: true,
-                ),
-              ],
-            ]),
-          );
-
-          try {
-            await channel.updateMessage(message, skipEnrichUrl: true);
-          } catch (e) {
-            expect(e, isA<StreamChatNetworkError>());
-
-            final networkError = e as StreamChatNetworkError;
-            expect(networkError.code, equals(ChatErrorCode.requestTimeout.code));
-            expect(networkError.isRetriable, isTrue);
-          }
-        },
-      );
-
-      test(
-        'should add message to retry queue when retriable StreamChatNetworkError occurs with skipPush: true, skipEnrichUrl: false',
-        () async {
-          final message = Message(
-            id: 'test-message-id-retry-2',
-            state: MessageState.sent,
-          );
-
-          // Create a retriable error (data == null)
-          when(
-            () => client.updateMessage(
-              any(that: isSameMessageAs(message)),
-              skipPush: true,
-            ),
-          ).thenThrow(
-            StreamChatNetworkError.raw(
-              code: ChatErrorCode.internalSystemError.code,
-              message: 'Internal system error',
-            ),
-          );
-
-          expectLater(
-            // skipping first seed message list -> [] messages
-            channel.state?.messagesStream.skip(1),
-            emitsInOrder([
-              [
-                isSameMessageAs(
-                  message.copyWith(state: MessageState.updating),
-                  matchMessageState: true,
-                ),
-              ],
-              [
-                isSameMessageAs(
-                  message.copyWith(
-                    state: MessageState.updatingFailed(
-                      skipPush: true,
-                      skipEnrichUrl: false,
-                    ),
-                  ),
-                  matchMessageState: true,
-                ),
-              ],
-            ]),
-          );
-
-          try {
-            await channel.updateMessage(message, skipPush: true);
-          } catch (e) {
-            expect(e, isA<StreamChatNetworkError>());
-
-            final networkError = e as StreamChatNetworkError;
-            expect(networkError.code, equals(ChatErrorCode.internalSystemError.code));
-            expect(networkError.isRetriable, isTrue);
-          }
-        },
-      );
-
-      test('should handle non-retriable StreamChatNetworkError with skipPush: true, skipEnrichUrl: true', () async {
-        final message = Message(
-          id: 'test-message-id-error-2',
-          state: MessageState.sent,
-        );
-
-        when(
-          () => client.updateMessage(
-            any(that: isSameMessageAs(message)),
-            skipPush: true,
-            skipEnrichUrl: true,
-          ),
-        ).thenThrow(StreamChatNetworkError(ChatErrorCode.notAllowed));
-
-        expectLater(
-          // skipping first seed message list -> [] messages
-          channel.state?.messagesStream.skip(1),
-          emitsInOrder([
-            [
-              isSameMessageAs(
-                message.copyWith(state: MessageState.updating),
-                matchMessageState: true,
-              ),
-            ],
-            [
-              isSameMessageAs(
-                message.copyWith(
-                  state: MessageState.updatingFailed(
-                    skipPush: true,
-                    skipEnrichUrl: true,
-                  ),
-                ),
-                matchMessageState: true,
-              ),
-            ],
-          ]),
-        );
-
-        try {
-          await channel.updateMessage(
-            message,
-            skipPush: true,
-            skipEnrichUrl: true,
-          );
-        } catch (e) {
-          expect(e, isA<StreamChatNetworkError>());
-
-          final networkError = e as StreamChatNetworkError;
-          expect(networkError.code, equals(ChatErrorCode.notAllowed.code));
-        }
-      });
-
-      test('should handle non-retriable StreamChatNetworkError with skipPush: false, skipEnrichUrl: false', () async {
-        final message = Message(
-          id: 'test-message-id-error-3',
-          state: MessageState.sent,
-        );
-
-        when(
-          () => client.updateMessage(
-            any(that: isSameMessageAs(message)),
-          ),
-        ).thenThrow(StreamChatNetworkError(ChatErrorCode.notAllowed));
-
-        expectLater(
-          // skipping first seed message list -> [] messages
-          channel.state?.messagesStream.skip(1),
-          emitsInOrder([
-            [
-              isSameMessageAs(
-                message.copyWith(state: MessageState.updating),
-                matchMessageState: true,
-              ),
-            ],
-            [
-              isSameMessageAs(
-                message.copyWith(
-                  state: MessageState.updatingFailed(
-                    skipPush: false,
-                    skipEnrichUrl: false,
-                  ),
-                ),
-                matchMessageState: true,
-              ),
-            ],
-          ]),
-        );
-
-        try {
-          await channel.updateMessage(message);
-        } catch (e) {
-          expect(e, isA<StreamChatNetworkError>());
-
-          final networkError = e as StreamChatNetworkError;
-          expect(networkError.code, equals(ChatErrorCode.notAllowed.code));
-        }
-      });
-    });
-
-    group('`ChannelClientState.updateMessage`', () {
-      test('upsert: true (default) adds an unknown message', () async {
-        final message = Message(
-          id: 'unknown-message',
-          user: client.state.currentUser,
-          text: 'hello',
-          createdAt: DateTime.utc(2026),
-        );
-
-        expect(channel.state!.messages, isEmpty);
-
-        channel.state!.updateMessage(message);
-
-        expect(channel.state!.messages.map((m) => m.id), ['unknown-message']);
-      });
-
-      test('upsert: false does NOT add an unknown message', () async {
-        final message = Message(
-          id: 'unknown-message',
-          user: client.state.currentUser,
-          text: 'hello',
-          createdAt: DateTime.utc(2026),
-        );
-
-        expect(channel.state!.messages, isEmpty);
-
-        channel.state!.updateMessage(message, upsert: false);
-
-        expect(channel.state!.messages, isEmpty);
-      });
-
-      test('upsert: false updates a message already in the window', () async {
-        const messageId = 'known-message';
-        final seeded = Message(
-          id: messageId,
-          user: client.state.currentUser,
-          text: 'old',
-          createdAt: DateTime.utc(2026),
-        );
-        channel.state!.updateChannelState(
-          channel.state!.channelState.copyWith(messages: [seeded]),
-        );
-
-        channel.state!.updateMessage(
-          seeded.copyWith(text: 'new'),
-          upsert: false,
-        );
-
-        final stored = channel.state!.messages.single;
-        expect(stored.id, equals(messageId));
-        expect(stored.text, equals('new'));
-      });
-    });
-
-    test('`.partialUpdateMessage`', () async {
-      final message = Message(
-        id: 'test-message-id',
-        state: MessageState.sent,
-      );
-
-      const set = {'text': 'Update Message text'};
-      const unset = ['pinExpires'];
-
-      final updateMessageResponse = UpdateMessageResponse()
-        ..message = message.copyWith(text: set['text'], pinExpires: null);
-
-      when(
-        () => client.partialUpdateMessage(message.id, set: set, unset: unset),
-      ).thenAnswer((_) async => updateMessageResponse);
-
-      expectLater(
-        // skipping first seed message list -> [] messages
-        channel.state?.messagesStream.skip(1),
-        emitsInOrder([
-          [
-            isSameMessageAs(
-              message.copyWith(
-                state: MessageState.updating,
-              ),
-              matchText: true,
-              matchMessageState: true,
-            ),
-          ],
-          [
-            isSameMessageAs(
-              updateMessageResponse.message.copyWith(
-                state: MessageState.updated,
-              ),
-              matchText: true,
-              matchMessageState: true,
-            ),
-          ],
-        ]),
-      );
-
-      final res = await channel.partialUpdateMessage(
-        message,
-        set: set,
-        unset: unset,
-      );
-
-      expect(res, isNotNull);
-      expect(res.message.id, message.id);
-      expect(res.message.id, message.id);
-      expect(res.message.text, set['text']);
-      expect(res.message.pinExpires, isNull);
-
-      verify(
-        () => client.partialUpdateMessage(message.id, set: set, unset: unset),
-      ).called(1);
-    });
-
-    group('`.partialUpdateMessage` error handling', () {
-      test('should update message state even when error is not StreamChatNetworkError', () async {
-        final message = Message(
-          id: 'test-message-id-error-partial-1',
-          state: MessageState.sent,
-        );
-
-        // Add message to channel state first
-        channel.state?.updateMessage(message);
-
-        const set = {'text': 'Update Message text'};
-        const unset = ['pinExpires'];
-
-        when(
-          () => client.partialUpdateMessage(
-            message.id,
-            set: set,
-            unset: unset,
-          ),
-        ).thenThrow(ArgumentError('Invalid argument'));
-
-        expectLater(
-          // skipping first seed message list -> [] messages
-          channel.state?.messagesStream.skip(1),
-          emitsInOrder([
-            [
-              isSameMessageAs(
-                message.copyWith(
-                  state: MessageState.updating,
-                ),
-                matchText: true,
-                matchMessageState: true,
-              ),
-            ],
-            [
-              isSameMessageAs(
-                message.copyWith(
-                  state: MessageState.partialUpdatingFailed(
-                    set: set,
-                    unset: unset,
-                    skipEnrichUrl: false,
-                  ),
-                ),
-                matchText: true,
-                matchMessageState: true,
-              ),
-            ],
-          ]),
-        );
-
-        try {
-          await channel.partialUpdateMessage(
-            message,
-            set: set,
-            unset: unset,
-          );
-        } catch (e) {
-          expect(e, isA<ArgumentError>());
-        }
-      });
-
-      test(
-        'should add message to retry queue when retriable StreamChatNetworkError occurs with skipEnrichUrl: true',
-        () async {
-          final message = Message(
-            id: 'test-message-id-retry-partial-1',
-            state: MessageState.sent,
-          );
-
-          // Add message to channel state first
-          channel.state?.updateMessage(message);
-
-          const set = {'text': 'Update Message text'};
-          const unset = ['pinExpires'];
-
-          // Create a retriable error (data == null)
-          when(
-            () => client.partialUpdateMessage(
-              message.id,
-              set: set,
-              unset: unset,
-              skipEnrichUrl: true,
-            ),
-          ).thenThrow(
-            StreamChatNetworkError.raw(
-              code: ChatErrorCode.requestTimeout.code,
-              message: 'Request timed out',
-            ),
-          );
-
-          expectLater(
-            // skipping first seed message list -> [] messages
-            channel.state?.messagesStream.skip(1),
-            emitsInOrder([
-              [
-                isSameMessageAs(
-                  message.copyWith(
-                    state: MessageState.updating,
-                  ),
-                  matchText: true,
-                  matchMessageState: true,
-                ),
-              ],
-              [
-                isSameMessageAs(
-                  message.copyWith(
-                    state: MessageState.partialUpdatingFailed(
-                      set: set,
-                      unset: unset,
-                      skipEnrichUrl: true,
-                    ),
-                  ),
-                  matchText: true,
-                  matchMessageState: true,
-                ),
-              ],
-            ]),
-          );
-
-          try {
-            await channel.partialUpdateMessage(
-              message,
-              set: set,
-              unset: unset,
-              skipEnrichUrl: true,
-            );
-          } catch (e) {
-            expect(e, isA<StreamChatNetworkError>());
-
-            final networkError = e as StreamChatNetworkError;
-            expect(networkError.code, equals(ChatErrorCode.requestTimeout.code));
-            expect(networkError.isRetriable, isTrue);
-          }
-        },
-      );
-
-      test(
-        'should add message to retry queue when retriable StreamChatNetworkError occurs with skipEnrichUrl: false',
-        () async {
-          final message = Message(
-            id: 'test-message-id-retry-partial-2',
-            state: MessageState.sent,
-          );
-
-          // Add message to channel state first
-          channel.state?.updateMessage(message);
-
-          const set = {'text': 'Update Message text'};
-          const unset = ['pinExpires'];
-
-          // Create a retriable error (data == null)
-          when(
-            () => client.partialUpdateMessage(
-              message.id,
-              set: set,
-              unset: unset,
-            ),
-          ).thenThrow(
-            StreamChatNetworkError.raw(
-              code: ChatErrorCode.internalSystemError.code,
-              message: 'Internal system error',
-            ),
-          );
-
-          expectLater(
-            // skipping first seed message list -> [] messages
-            channel.state?.messagesStream.skip(1),
-            emitsInOrder([
-              [
-                isSameMessageAs(
-                  message.copyWith(
-                    state: MessageState.updating,
-                  ),
-                  matchText: true,
-                  matchMessageState: true,
-                ),
-              ],
-              [
-                isSameMessageAs(
-                  message.copyWith(
-                    state: MessageState.partialUpdatingFailed(
-                      set: set,
-                      unset: unset,
-                      skipEnrichUrl: false,
-                    ),
-                  ),
-                  matchText: true,
-                  matchMessageState: true,
-                ),
-              ],
-            ]),
-          );
-
-          try {
-            await channel.partialUpdateMessage(
-              message,
-              set: set,
-              unset: unset,
-            );
-          } catch (e) {
-            expect(e, isA<StreamChatNetworkError>());
-
-            final networkError = e as StreamChatNetworkError;
-            expect(networkError.code, equals(ChatErrorCode.internalSystemError.code));
-            expect(networkError.isRetriable, isTrue);
-          }
-        },
-      );
-
-      test('should handle non-retriable StreamChatNetworkError with skipEnrichUrl: true', () async {
-        final message = Message(
-          id: 'test-message-id-error-partial-2',
-          state: MessageState.sent,
-        );
-
-        // Add message to channel state first
-        channel.state?.updateMessage(message);
-
-        const set = {'text': 'Update Message text'};
-        const unset = ['pinExpires'];
-
-        when(
-          () => client.partialUpdateMessage(
-            message.id,
-            set: set,
-            unset: unset,
-            skipEnrichUrl: true,
-          ),
-        ).thenThrow(StreamChatNetworkError(ChatErrorCode.notAllowed));
-
-        expectLater(
-          // skipping first seed message list -> [] messages
-          channel.state?.messagesStream.skip(1),
-          emitsInOrder([
-            [
-              isSameMessageAs(
-                message.copyWith(
-                  state: MessageState.updating,
-                ),
-                matchText: true,
-                matchMessageState: true,
-              ),
-            ],
-            [
-              isSameMessageAs(
-                message.copyWith(
-                  state: MessageState.partialUpdatingFailed(
-                    set: set,
-                    unset: unset,
-                    skipEnrichUrl: true,
-                  ),
-                ),
-                matchText: true,
-                matchMessageState: true,
-              ),
-            ],
-          ]),
-        );
-
-        try {
-          await channel.partialUpdateMessage(
-            message,
-            set: set,
-            unset: unset,
-            skipEnrichUrl: true,
-          );
-        } catch (e) {
-          expect(e, isA<StreamChatNetworkError>());
-
-          final networkError = e as StreamChatNetworkError;
-          expect(networkError.code, equals(ChatErrorCode.notAllowed.code));
-        }
-      });
-
-      test('should handle non-retriable StreamChatNetworkError with skipEnrichUrl: false', () async {
-        final message = Message(
-          id: 'test-message-id-error-partial-3',
-          state: MessageState.sent,
-        );
-
-        // Add message to channel state first
-        channel.state?.updateMessage(message);
-
-        const set = {'text': 'Update Message text'};
-        const unset = ['pinExpires'];
-
-        when(
-          () => client.partialUpdateMessage(
-            message.id,
-            set: set,
-            unset: unset,
-          ),
-        ).thenThrow(StreamChatNetworkError(ChatErrorCode.notAllowed));
-
-        expectLater(
-          // skipping first seed message list -> [] messages
-          channel.state?.messagesStream.skip(1),
-          emitsInOrder([
-            [
-              isSameMessageAs(
-                message.copyWith(
-                  state: MessageState.updating,
-                ),
-                matchText: true,
-                matchMessageState: true,
-              ),
-            ],
-            [
-              isSameMessageAs(
-                message.copyWith(
-                  state: MessageState.partialUpdatingFailed(
-                    set: set,
-                    unset: unset,
-                    skipEnrichUrl: false,
-                  ),
-                ),
-                matchText: true,
-                matchMessageState: true,
-              ),
-            ],
-          ]),
-        );
-
-        try {
-          await channel.partialUpdateMessage(
-            message,
-            set: set,
-            unset: unset,
-          );
-        } catch (e) {
-          expect(e, isA<StreamChatNetworkError>());
-
-          final networkError = e as StreamChatNetworkError;
-          expect(networkError.code, equals(ChatErrorCode.notAllowed.code));
-        }
-      });
-    });
-
     group('`.sendReaction`', () {
       test('should work fine', () async {
         final message = Message(
